+++ internal/controller/operator/vmuser_controller.go
@@ -116,7 +116,7 @@ func (r *VMUserReconciler) Reconcile(ctx context.Context, req ctrl.Request) (res
 				continue
 			}
 		}
-		if err := vmauth.CreateOrUpdateVMAuthConfig(ctx, r, currentVMAuth); err != nil {
+		if err := vmauth.CreateOrUpdateVMAuthConfig(ctx, r, currentVMAuth, &instance); err != nil {
 			return ctrl.Result{}, fmt.Errorf("cannot create or update vmauth deploy for vmuser: %w", err)
 		}
 	}