+++ internal/controller/operator/factory/reconcile/deploy.go
@@ -49,9 +49,10 @@ func Deployment(ctx context.Context, rclient client.Client, newDeploy, prevDeplo
 			newDeploy.Spec.Replicas = currentDeploy.Spec.Replicas
 		}
 		newDeploy.Status = currentDeploy.Status
-		var prevAnnotations map[string]string
+		var prevAnnotations, prevTemplateAnnotations map[string]string
 		if prevDeploy != nil {
 			prevAnnotations = prevDeploy.Annotations
+			prevTemplateAnnotations = prevDeploy.Spec.Template.Annotations
 		}
 		isEqual := equality.Semantic.DeepDerivative(newDeploy.Spec, currentDeploy.Spec)
 		if isEqual &&
@@ -63,6 +64,7 @@ func Deployment(ctx context.Context, rclient client.Client, newDeploy, prevDeplo
 
 		vmv1beta1.AddFinalizer(newDeploy, &currentDeploy)
 		newDeploy.Annotations = mergeAnnotations(currentDeploy.Annotations, newDeploy.Annotations, prevAnnotations)
+		newDeploy.Spec.Template.Annotations = mergeAnnotations(currentDeploy.Spec.Template.Annotations, newDeploy.Spec.Template.Annotations, prevTemplateAnnotations)
 		cloneSignificantMetadata(newDeploy, &currentDeploy)
 
 		logger.WithContext(ctx).Info(fmt.Sprintf("updating Deployment %s configuration"+